Migrating affix to MUI
